+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/entity/MenuItem.kt
@@ -20,5 +20,5 @@ data class MenuItem (
         var ingredients: List<Ingredient>,
 
         @Column
-        var price: Long
+        var price: Double
 )
